mv keychain to its own package
